alura-django-rest-customer/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-filter==22.1
markdown==3.4.1
asgiref==3.5.2
astroid==2.12.2
lazy-object-proxy==1.7.1
mccabe==0.7.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.5.4
wrapt==1.14.1
validate-docbr==1.10.0
Faker==14.1.0
isort
black
